Remove requireAuthentication from restmap.conf
Unfortunately, we cannot require authentication in the custom endpoint,
because any unauthenticated request would get sent back either HTML or
XML, and we wouldn't be able to intercept it.  The problem with not
using requireAuthentication is that we can no longer "log in" with basic
auth, which means we can't use it in the browser.