Add update functionality to Entity, change remove to callback with nothing, and add some tests
This change adds the ability to update an entity with a dictionary of
values (which will get sent to the server, and invalidate the object).
It adds tests that update an app and check to see that the updates took.

It also fixes a "bug" where calling "remove" would return the response
from the server, even though it is not necessary. This makes it easier
to chain it.

Finally, I added a "cleanup" test to remove all apps created by the
test suite.