Fix a case where we would die due to exec_mode=oneshot returning bad XML
If you make a search request with exec_mode=oneshot, and the search has
no results, then you would simply get back "<?xml version='1.0'
encoding='UTF-8'?>", which ElementTree cannot parse. We now detect this
case and replace it with an empty results tag. Also, a test is added to
catch this case.