Fix `job.track()` to handle new 'ready' check (DVPL-3555)
In Splunk 6, we can now get content when the job is not ready yet,. As
such, we need to check the dispatch state to note when the job is
ready, rather than relying on the absence of data to decide it isn't
ready.

This change also fixes the setpriority tests to use `job.track`.